+++ src/components/TiptapDirectionsAdd.jsx
@@ -15,32 +15,9 @@ import { FormatBoldOutlined, FormatItalicOutlined, FormatUnderlinedOutlined, Str
   UndoOutlined, RedoOutlined} from '@mui/icons-material'
 import { Box} from '@mui/material';
 import MenuBar from './TiptapMenuBar'
+import extensions from './TiptapExtensions'
 
-
-// define your extension array
-const extensions = [
-  Color.configure({ types: [TextStyle.name, ListItem.name] }),
-  TextStyle.configure({ types: [ListItem.name] }),
-  StarterKit.configure({
-    bulletList: {
-      keepMarks: true,
-      keepAttributes: false, // TODO : Making this as `false` becase marks are not preserved when I try to preserve attrs, awaiting a bit of help
-    },
-    orderedList: {
-      keepMarks: true,
-      keepAttributes: false, // TODO : Making this as `false` becase marks are not preserved when I try to preserve attrs, awaiting a bit of help
-    }
-  }),
-  Placeholder.configure({
-    placeholder: 'Directions',
-  }),
-  Underline, 
-  TextAlign.configure({
-    types: ['heading', 'paragraph'], 
-    alignments: ['left', 'right'],
-    defaultAlignment: 'left',
-  })
-]
+< extensions/>
 
 const content = ''
