+++ src/components/TiptapIngredientsAdd.jsx
@@ -13,42 +13,9 @@ import Underline from '@tiptap/extension-underline'
 import TextAlign from '@tiptap/extension-text-align'
 import { Box} from '@mui/material';
 import MenuBar from './TiptapMenuBar'
-import { TurnLeft } from '@mui/icons-material'
+import extensions from './TiptapExtensions'
 
-  
-// define your extension array
-const extensions = [
-  Color.configure({ types: [TextStyle.name, ListItem.name] }),
-  TextStyle.configure({ types: [ListItem.name] }),
-  StarterKit.configure({
-    bulletList: {
-      keepMarks: true,
-      keepAttributes: false, // TODO : Making this as `false` becase marks are not preserved when I try to preserve attrs, awaiting a bit of help
-    },
-    orderedList: {
-      keepMarks: true,
-      keepAttributes: false, // TODO : Making this as `false` becase marks are not preserved when I try to preserve attrs, awaiting a bit of help
-    }
-  }),
-  Placeholder.configure({
-    placeholder: 'Ingredients',
-  }),
-  Underline, 
-  TextAlign.configure({
-    types: ['heading', 'paragraph'], 
-    alignments: ['left', 'right'],
-    defaultAlignment: 'left',
-  }),
-  TaskList.configure({
-    HTMLAttributes: {
-      class: 'taskStyle',
-    }
-  }),
-
-  TaskItem.configure({
-    nested: true,
-  })
-]
+< extensions/>
 
 const content = ''
